+++ inst/examples/bru.R
@@ -1,9 +1,10 @@
 \donttest{
 if (bru_safe_inla()) {
-
   # Simulate some covariates x and observations y
   input.df <- data.frame(x = cos(1:10))
-  input.df <- within(input.df, y <- 5 + 2 * x + rnorm(10, mean = 0, sd = 0.1))
+  input.df <- within(input.df, {
+    y <- 5 + 2 * x + rnorm(10, mean = 0, sd = 0.1)
+  })
 
   # Fit a Gaussian likelihood model
   fit <- bru(y ~ x + Intercept(1), family = "gaussian", data = input.df)
@@ -14,10 +15,11 @@ if (bru_safe_inla()) {
 
 
 if (bru_safe_inla()) {
-
   # Alternatively, we can use the like() function to construct the likelihood:
 
-  lik <- like(family = "gaussian", formula = y ~ x + Intercept, data = input.df)
+  lik <- like(family = "gaussian",
+              formula = y ~ x + Intercept,
+              data = input.df)
   fit <- bru(~ x + Intercept(1), lik)
   fit$summary.fixed
 }
@@ -30,7 +32,9 @@ if (bru_safe_inla()) {
 
 if (bru_safe_inla()) {
   z <- 2
-  input.df <- within(input.df, y <- 5 + exp(z) * x + rnorm(10, mean = 0, sd = 0.1))
+  input.df <- within(input.df, {
+    y <- 5 + exp(z) * x + rnorm(10, mean = 0, sd = 0.1)
+  })
   lik <- like(
     family = "gaussian", data = input.df,
     formula = y ~ exp(z) * x + Intercept