+++ NEWS.md
@@ -7,6 +7,7 @@
   as input require either `sp::coordinates` or `sp` having been already loaded
   with e.g. `library(sp)` (version `2.11.1.9003`)
 * Use `bru_safe_inla(multicore=TRUE)` in `bru()` to allow multicore in non-interactive mode. Fixes regression bug that disabled multicore for non-interactive sessions. (version `2.11.1.9004`)
+* Reduced the amount of diagnostic messages in `bru_safe_inla()` (version `2.11.1.9005`)
 
 # inlabru 2.11.1